+++ 05-react-context/src/components/HeaderContext.js
@@ -1,9 +1,13 @@
 import { useContext } from 'react'
+import AuthContext from '../context/AuthContext'
+import LanguageContext from '../context/LanguageContext'
 import ThemeContext from '../context/ThemeConrtext'
 
-const HeaderContext = ({ texts, handleLanguage, auth, handleAuth }) => {
+const HeaderContext = () => {
   // Destructuramos las variables del contextop creado
   const { theme, handleTheme } = useContext(ThemeContext)
+  const { texts, handleLanguage } = useContext(LanguageContext)
+  const { auth, handleAuth } = useContext(AuthContext)
 
   return (
     <header className={theme}>
