+++ 05-react-context/src/context/AuthContext.js
@@ -1,0 +1,25 @@
+import { createContext, useState } from 'react'
+
+const AuthContext = createContext()
+
+const initialAuth = null
+
+const AuthProvider = ({ children }) => {
+  const [auth, setAuth] = useState(initialAuth)
+
+  // Sesion change handling
+  const handleAuth = (e) => {
+    if (auth) {
+      setAuth(null)
+    } else {
+      setAuth(true)
+    }
+  }
+
+  const data = { auth, handleAuth }
+
+  return <AuthContext.Provider value={data}>{children}</AuthContext.Provider>
+}
+
+export { AuthProvider }
+export default AuthContext
